+++ client/game/Empire/Empire.vue
@@ -37,18 +37,3 @@
         }
     }
 </style>
-
-<style lang="scss">
-    .v--modal-overlay {
-        background: rgba(palette("grey", "black"), 0.5);
-    }
-
-    .v--modal {
-        border: 3px solid palette("grey", "charcoal");
-
-        background: palette("grey", "sunken");
-        color: palette("text");
-        border-radius: 0;
-        box-shadow: 0 0 60px rgba(palette("grey", "mystic"), 0.15);
-    }
-</style>