+++ server/config/game/planets.js
@@ -5,6 +5,7 @@
  * they do not necessarily add up to 100.
  * the resourceSlot chances, however, are percentage chances. they are also cumulative:
  * if a 30 is rolled and the chance was 70, there is a new chance with 40% - up to max prop
+ * potential is the "grade" of the resource;
  *
  **********************************************************************************************************************/
 const RULES =  {
@@ -14,8 +15,8 @@ const RULES =  {
             chance: 20,
             chanceHome: 37,
             resourceSlots: [
-                {type: "food", chance: 70, max: 3, potential: [3000, 6000]},
-                {type: "energy", chance: 50, max: 2, potential: [2000, 3000]}
+                {type: "food", chance: 70, max: 3, potential: [1, 2]},
+                {type: "energy", chance: 50, max: 2, potential: [0.7, 1.3]}
             ]
         },
         {
@@ -23,8 +24,8 @@ const RULES =  {
             chance: 17,
             chanceHome: 13,
             resourceSlots: [
-                {type: "research", chance: 70, max: 3, potential: [3000, 6000]},
-                {type: "energy", chance: 50, max: 2, potential: [2000, 3000]}
+                {type: "research", chance: 70, max: 3, potential: [1, 2]},
+                {type: "energy", chance: 50, max: 2, potential: [0.7, 1.3]}
             ]
         },
         {
@@ -32,9 +33,9 @@ const RULES =  {
             chance: 16,
             chanceHome: 10,
             resourceSlots: [
-                {type: "minerals", chance: 30, max: 1, potential: [1200, 2400]},
-                {type: "food", chance: 70, max: 2, potential: [3000, 6000]},
-                {type: "research", chance: 15, max: 1, potential: [750, 1500]}
+                {type: "minerals", chance: 30, max: 1, potential: [0.6, 1.2]},
+                {type: "food", chance: 70, max: 2, potential: [1, 2]},
+                {type: "research", chance: 15, max: 1, potential: [0.5, 1]}
             ]
         },
         {
@@ -42,9 +43,9 @@ const RULES =  {
             chance: 20,
             chanceHome: 10,
             resourceSlots: [
-                {type: "minerals", chance: 70, max: 3, potential: [3000, 6000]},
-                {type: "research", chance: 30, max: 1, potential: [1200, 2400]},
-                {type: "food", chance: 15, max: 1, potential: [750, 1500]}
+                {type: "minerals", chance: 70, max: 3, potential: [1, 2]},
+                {type: "research", chance: 30, max: 1, potential: [0.6, 1.2]},
+                {type: "food", chance: 15, max: 1, potential: [0.5, 1]}
             ]
         },
         {
@@ -52,9 +53,9 @@ const RULES =  {
             chance: 19,
             chanceHome: 10,
             resourceSlots: [
-                {type: "energy", chance: 50, max: 2, potential: [2000, 3000]},
-                {type: "food", chance: 50, max: 2, potential: [2000, 3000]},
-                {type: "minerals", chance: 15, max: 1, potential: [750, 1500]}
+                {type: "energy", chance: 50, max: 2, potential: [0.7, 1.3]},
+                {type: "food", chance: 50, max: 2, potential: [0.7, 1.3]},
+                {type: "minerals", chance: 15, max: 1, potential: [0.5, 1]}
             ]
         },
         {
@@ -62,9 +63,9 @@ const RULES =  {
             chance: 10,
             chanceHome: 13,
             resourceSlots: [
-                {type: "minerals", chance: 50, max: 2, potential: [2000, 3000]},
-                {type: "energy", chance: 50, max: 2, potential: [2000, 3000]},
-                {type: "research", chance: 15, max: 1, potential: [750, 1500]}
+                {type: "minerals", chance: 50, max: 2, potential: [0.7, 1.3]},
+                {type: "energy", chance: 50, max: 2, potential: [0.7, 1.3]},
+                {type: "research", chance: 15, max: 1, potential: [0.5, 1]}
             ]
         },
         {
@@ -72,8 +73,8 @@ const RULES =  {
             chance: 7,
             chanceHome: 7,
             resourceSlots: [
-                {type: "minerals", chance: 50, max: 2, potential: [2000, 3000]},
-                {type: "research", chance: 70, max: 3, potential: [3000, 6000]}
+                {type: "minerals", chance: 50, max: 2, potential: [0.7, 1.3]},
+                {type: "research", chance: 70, max: 3, potential: [1, 2]}
             ]
         },
         {
@@ -81,10 +82,10 @@ const RULES =  {
             chance: 1,
             chanceHome: 0,
             resourceSlots: [
-                {type: "energy", chance: 30, max: 1, potential: [1000, 2000]},
-                {type: "minerals", chance: 30, max: 1, potential: [1000, 2000]},
-                {type: "food", chance: 30, max: 1, potential: [1000, 2000]},
-                {type: "research", chance: 100, max: 3, potential: [4000, 8000]}
+                {type: "energy", chance: 30, max: 1, potential: [0.6, 1.2]},
+                {type: "minerals", chance: 30, max: 1, potential: [0.6, 1.2]},
+                {type: "food", chance: 30, max: 1, potential: [0.6, 1.2]},
+                {type: "research", chance: 100, max: 3, potential: [1.2, 2.4]}
             ]
         }
     ]